requirements.txt
mcrcon
python-telegram-bot<=13.15
requests